+++ src/main/java/com/library/libraryapi/DAO/AddressRepository.java
@@ -1,8 +1,8 @@
 package com.library.libraryapi.DAO;
 
-import com.library.libraryapi.Model.AddressModel;
+import com.library.libraryapi.Model.Address;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AddressRepository extends CrudRepository<AddressModel, Long> {
+public interface AddressRepository extends CrudRepository<Address, Long> {
 
 }