+++ src/main/java/com/library/libraryapi/DAO/CustomerRepository.java
@@ -1,9 +1,9 @@
 package com.library.libraryapi.DAO;
 
-import com.library.libraryapi.Model.CustomerModel;
+import com.library.libraryapi.Model.Customer;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<CustomerModel, Long> {
-    CustomerModel findByLoginAndPassword(String login, String password);
-    CustomerModel findByLogin(String login);
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
+    Customer findByLoginAndPassword(String login, String password);
+    Customer findByLogin(String login);
 }