+++ src/main/java/com/library/libraryapi/Model/DictionaryItem.java
@@ -1,0 +1,66 @@
+package com.library.libraryapi.Model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
+
+@Entity
+public class DictionaryItem {
+
+    @Id
+    @GeneratedValue
+    @NotNull
+    private long id;
+
+    @NotNull
+    private String name;
+
+    @NotNull
+    private String code;
+
+    @NotNull
+    private String domain;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    public String getDomain() {
+        return domain;
+    }
+
+    public void setDomain(String domain) {
+        this.domain = domain;
+    }
+
+    @Override
+    public String toString() {
+        return "DictionaryItem{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", code='" + code + '\'' +
+                ", domain='" + domain + '\'' +
+                '}';
+    }
+}